requirements.txt
# Google APIs
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-generativeai==0.7.2
google-cloud-texttospeech==2.16.3
google-cloud-storage==2.16.0
google-api-python-client==2.130.0

# 動画・音声処理
moviepy==1.0.3
openai-whisper==20231117
ffmpeg-python==0.2.0

# HTTP
requests==2.32.3
